Rename `crate::yubikey` to `crate::key`
So that it doesn't conflict with the renamed `yubikey` crate.
